+++ casepro/rules/tests.py
@@ -19,6 +19,7 @@ from .models import (
     Test,
     WordCountTest,
 )
+from .templatetags.rules import render_actions, render_tests
 
 
 class TestsTest(BaseCasesTest):
@@ -131,6 +132,62 @@ class TestsTest(BaseCasesTest):
         self.assertFalse(ContainsTest.is_valid_keyword("-kit"))  # can't start with a dash
         self.assertFalse(ContainsTest.is_valid_keyword("kat "))  # can't end with a space
         self.assertFalse(ContainsTest.is_valid_keyword("kat-"))  # can't end with a dash
+
+
+class RulesTemplateTagsTest(BaseCasesTest):
+    def setUp(self):
+        super(RulesTemplateTagsTest, self).setUp()
+
+        self.context = DeserializationContext(self.unicef)
+
+    def test_render_contains_test(self):
+        # single keyword
+        test = Test.from_json({"type": "contains", "keywords": ["blue"], "quantifier": "all"}, self.context)
+        rule = Rule.create(self.unicef, [test], [])
+        self.assertEqual('message contains <i>"blue"</i>', render_tests(rule))
+
+        # multiple keywords
+        test = Test.from_json({"type": "contains", "keywords": ["RED", "Blue"], "quantifier": "any"}, self.context)
+        rule = Rule.create(self.unicef, [test], [])
+        self.assertEqual('message contains any of <i>"red"</i>, <i>"blue"</i>', render_tests(rule))
+
+    def test_render_word_count_test(self):
+        test = Test.from_json({"type": "words", "minimum": 2}, self.context)
+        rule = Rule.create(self.unicef, [test], [])
+        self.assertEqual("message has at least 2 words", render_tests(rule))
+
+    def test_render_group_test(self):
+        # single group
+        test = Test.from_json({"type": "groups", "groups": [self.reporters.pk], "quantifier": "any"}, self.context)
+        rule = Rule.create(self.unicef, [test], [])
+        self.assertEqual("contact belongs to <i>Reporters</i>", render_tests(rule))
+
+        # multiple groups
+        test = Test.from_json(
+            {"type": "groups", "groups": [self.females.pk, self.reporters.pk], "quantifier": "any"}, self.context
+        )
+        rule = Rule.create(self.unicef, [test], [])
+        self.assertEqual("contact belongs to any of <i>Females</i>, <i>Reporters</i>", render_tests(rule))
+
+    def test_render_field_test(self):
+        test = Test.from_json({"type": "field", "key": "city", "values": ["Kigali"]}, self.context)
+        rule = Rule.create(self.unicef, [test], [])
+        self.assertEqual("contact <i>city</i> is equal to <i>kigali</i>", render_tests(rule))
+
+    def test_render_label_action(self):
+        action = Action.from_json({"type": "label", "label": self.aids.pk}, self.context)
+        rule = Rule.create(self.unicef, [], [action])
+        self.assertEqual('apply label <span class="label label-success">AIDS</span>', render_actions(rule))
+
+    def test_render_flag_action(self):
+        action = Action.from_json({"type": "flag"}, self.context)
+        rule = Rule.create(self.unicef, [], [action])
+        self.assertEqual("flag message", render_actions(rule))
+
+    def test_render_archive_action(self):
+        action = Action.from_json({"type": "archive"}, self.context)
+        rule = Rule.create(self.unicef, [], [action])
+        self.assertEqual("archive message", render_actions(rule))
 
 
 class ActionsTest(BaseCasesTest):